Dax kmem api extracted to header files
