+++ copying_headers/MANIFEST.freeBSD
@@ -122,6 +122,8 @@ test/bat_tests.cpp
 test/check.cpp
 test/check.h
 test/common.h
+test/dax_kmem_nodes.cpp
+test/dax_kmem_nodes.h
 test/dax_kmem_env_var_test.py
 test/decorator_test.cpp
 test/decorator_test.h
