+++ test/TestPolicy.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 - 2019 Intel Corporation.
+ * Copyright (C) 2017 - 2020 Intel Corporation.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -155,5 +155,23 @@ namespace TestPolicy
         }
 
         check_numa_nodes(expected_bitmask, policy, ptr, size);
+    }
+
+    std::set<int> get_regular_numa_nodes(void)
+    {
+        struct bitmask *cpu_mask = numa_allocate_cpumask();
+        std::set<int> regular_nodes;
+
+        const int MAXNODE_ID = numa_max_node();
+        for (int id = 0; id <= MAXNODE_ID; ++id) {
+            numa_node_to_cpus(id, cpu_mask);
+
+            if (numa_bitmask_weight(cpu_mask) != 0) {
+                regular_nodes.insert(id);
+            }
+        }
+        numa_free_cpumask(cpu_mask);
+
+        return regular_nodes;
     }
 }
