+++ test/dax_kmem_nodes.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 - 2020 Intel Corporation.
+ * Copyright (C) 2020 Intel Corporation.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,53 +22,18 @@
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#pragma once
-#include <cstring>
-#include <fstream>
+#include <set>
 
-class ProcStat
+class DaxKmem
 {
 public:
-    size_t get_virtual_memory_size_bytes()
-    {
-        get_stat("VmSize", str_value);
-        return strtol(str_value, NULL, 10) * 1024;
-    }
+    DaxKmem();
+    size_t get_free_space(void);
+    std::set<int> get_closest_numa_nodes(int regular_node);
+    size_t size(void);
+    bool contains(int node);
 
-    size_t get_physical_memory_size_bytes()
-    {
-        get_stat("VmRSS", str_value);
-        return strtol(str_value, NULL, 10) * 1024;
-    }
-
-    size_t get_used_swap_space_size_bytes(void)
-    {
-        get_stat("VmSwap", str_value);
-        return strtol(str_value, NULL, 10) * 1024;
-    }
 private:
-    /* We are avoiding to allocate local buffers,
-     * since it can produce noise in memory footprint tests.
-     */
-    char line[1024];
-    char current_entry_name[1024];
-    char str_value[1024];
-
-    // Note: this function is not thread-safe.
-    void get_stat(const char *field_name, char *value)
-    {
-        char *pos = nullptr;
-        std::ifstream file("/proc/self/status", std::ifstream::in);
-        if (file.is_open()) {
-            while (file.getline(line, sizeof(line))) {
-                pos = strstr(line, field_name);
-                if (pos) {
-                    sscanf(pos, "%64[a-zA-Z_0-9()]: %s", current_entry_name, value);
-                    break;
-                }
-            }
-            file.close();
-        }
-    }
-};
-
+    std::set<int> dax_kmem_nodes;
+    std::set<int> get_dax_kmem_nodes(void);
+}; // DaxKmem
